Fixes a bug where a CONSTRUCT + UNION were resulting in extra bnodes with an id of "LL" being generated.
Function handle_construct checks for bnodes with a value of "NULL" instead of a bnode id, and skips outputting them.

Function slot_fill sets lexical value of bnode to b->lex rather than (b->lex)+2 when it encounters a value of "NULL" instead of a bnode id.

Added check that row isn't NULL before checking row->stop in a number of functions.