Small patch to stop make failing on Ubuntu 12.04 (and others)
On Ubuntu 12.04 (and probably others), make fails with "undefined reference to `uuid_generate'" errors while in src/http. Checking the make log, it appears that the -luuid argument was not included in the libtool steps. Adding the @UUID_LIBS@ arguments to the LDADD steps in the http/Makefile.am seems to set things straight.