Compile on OpenBSD, use different variants of libuuid
This is a minimal set of changes needed to allow 4store to
compile on OpenBSD successfully. It also removes the hard
dependency on Linux-style libuuid by detecting and coping
with OSSP or BSD-style libuuids as well.

Note that the now-successful build results in binaries that
still don't quite work properly, because of differences in
handling IPv6 between OpenBSD and Linux. These will be
the subject of a separate patch.