back out nonessential changes from 2b2338b
Commit 2b2338b changed the --with-storage-path switch of the configure script
to --with-default-storage-path. That change is not necessary to enable the user
to change the store root at runtime, and breaks some existing installations.
This commit puts the old name back.

Commit 2b2338b changed the FS_STORE_ROOT autoconf substitution variable to
FS_DEFAULT_STORE_ROOT. That change is not really necessary either, and broke
src/utilities/Makefile.am. This commit puts the old name back.